eng-fluency-platform/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
email-validator>=2.0.0
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
redis==5.0.3
celery==5.3.6
pandas==2.2.1
openpyxl==3.1.2
websockets==12.0
# openai removed
pydub==0.25.1
# langchain-openai removed
httpx==0.27.0
azure-cognitiveservices-speech==1.35.0
numpy==1.26.4
scipy==1.12.0
reportlab==4.0.9
stripe==8.5.0
google-generativeai
langchain-google-genai
